requirements.txt
anyio==3.6.2
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgiref==3.7.2
asttokens==2.2.1
attrs==22.2.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
captcha==0.4
cffi==1.15.1
click==8.1.3
colorama==0.4.6
comm==0.1.2
cssbeautifier==1.14.11
debugpy==1.6.4
decorator==5.1.1
defusedxml==0.7.1
Django==4.2.7
djangorestframework==3.14.0
EditorConfig==0.12.3
entrypoints==0.4
executing==1.2.0
fastjsonschema==2.16.2
Flask==2.2.2
Flask-Cors==3.0.10
fqdn==1.5.1
html-tag-names==0.1.2
html-void-elements==0.1.0
idna==3.4
inflection==0.5.1
ipykernel==6.20.0
ipython==8.7.0
ipython-genutils==0.2.0
ipywidgets==8.0.4
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
jsbeautifier==1.14.11
json5==0.9.14
jsonify==0.5
jsonpointer==2.3
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.4.4
jupyter-events==0.5.0
jupyter_client==7.4.8
jupyter_core==5.1.1
jupyter_server==2.0.5
jupyter_server_terminals==0.4.3
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.5
MarkupSafe==2.1.1
matplotlib-inline==0.1.6
mistune==2.0.4
nbclassic==0.4.8
nbclient==0.7.2
nbconvert==7.2.7
nbformat==5.7.1
nest-asyncio==1.5.6
notebook==6.5.2
notebook_shim==0.2.2
packaging==22.0
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.11.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.3.0
platformdirs==2.6.0
prometheus-client==0.15.0
prompt-toolkit==3.0.36
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.13.0
pyrsistent==0.19.2
python-dateutil==2.8.2
python-json-logger==2.0.4
pytz==2023.3.post1
PyYAML==6.0
pyzmq==24.0.1
qtconsole==5.4.0
QtPy==2.3.0
regex==2023.10.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
Send2Trash==1.8.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.3.2.post1
sqlparse==0.4.4
stack-data==0.6.2
terminado==0.17.1
tinycss2==1.2.1
tornado==6.2
tqdm==4.66.1
traitlets==5.8.0
uri-template==1.2.0
uritemplate==4.1.1
wcwidth==0.2.5
webcolors==1.12
webencodings==0.5.1
websocket-client==1.4.2
Werkzeug==2.2.2
widgetsnbextension==4.0.5
